feat: add editor sidebar and device chooser

+++ core/preview/src/main/kotlin/com/avin/avinapp/device/PreviewDevice.kt
@@ -7,7 +7,7 @@ data class PreviewDevice(
     val name: String,
     val resolution: Resolution,
     val type: DeviceType,
-    val density: Double
+    val density: Float
 ) {
     @Serializable
     data class Resolution(
